+++ NEWS.md
@@ -12,6 +12,7 @@
 * `sg_change_*_p` family added to change nodes and edges attributes on the fly.
 * Event improved.
 * `sg_read_*_p` family of functions to easily add nodes and edges in bulk.
+* `sg_read_batch_*_p` family of functions to easily add nodes and edges in bulk and by batch.
 
 # sigmajs 0.1.2